// src/app/modelos/estudiante.interface.ts
export interface EstudianteI{
   id_estudiante: string;
   codigo: string;
   dni: string;
   nombre: string;
   apellido: string;
   escuela: string;
   facultad: string;
   nrocelular: string;
   direccion: string;
   sexo: string;
   correo: string;
   //
   estado: string,
   nro_hito: string
   id_tramite
}